+++ kai/ukernels/matmul/matmul_clamp_f16_f16_f16p/kai_matmul_clamp_f16_f16_f16p2vlx2b_1x16vl_sme2_dot.h
@@ -107,9 +107,13 @@ size_t kai_get_dst_size_matmul_clamp_f16_f16_f16p2vlx2b_1x16vl_sme2_dot(size_t m
 /// @param[in]  dst_stride_col Column stride in bytes of the output matrix. Currently, an unused parameter.
 /// @param[in]  clamp_min Minimum value to clamp the final result.
 /// @param[in]  clamp_max Maximum value to clamp the final result.
+///
+/// @note Clamp minimum and maximum values are cast internally to the destination type before clamping the computed
+/// values.
+///
 void kai_run_matmul_clamp_f16_f16_f16p2vlx2b_1x16vl_sme2_dot(
     size_t m, size_t n, size_t k, const void* lhs, size_t lhs_stride, const void* rhs_packed, void* dst,
-    size_t dst_stride_row, size_t dst_stride_col, __fp16 clamp_min, __fp16 clamp_max);
+    size_t dst_stride_row, size_t dst_stride_col, float clamp_min, float clamp_max);
 
 #ifdef __cplusplus
 }  // extern "C"
