+++ test/common/matmul_test_common.hpp
@@ -297,7 +297,7 @@ struct MatMulMethod {
         const void* lhs, size_t lhs_stride,                       //
         const void* packed_rhs,                                   //
         void* dst, size_t dst_stride_row, size_t dst_stride_col,  //
-        __fp16 clamp_min, __fp16 clamp_max)>
+        float clamp_min, float clamp_max)>
         fn_matmul_f16_f16_f16p = nullptr;
 
     std::function<void(
@@ -433,12 +433,9 @@ struct MatMulMethod {
 
         if (fn_matmul_f16_f16_f16p) {
             fn_matmul_f16_f16_f16p(
-                m, n, k, lhs, lhs_stride, rhs, dst, dst_stride, sizeof(Float16), clamp_min,
-                static_cast<__fp16>(clamp_max));
+                m, n, k, lhs, lhs_stride, rhs, dst, dst_stride, sizeof(uint16_t), clamp_min, clamp_max);
         } else if (fn_matmul_f32_f32_f32p) {
-            fn_matmul_f32_f32_f32p(
-                m, n, k, lhs, lhs_stride, rhs, dst, dst_stride, sizeof(float), clamp_min,
-                static_cast<__fp16>(clamp_max));
+            fn_matmul_f32_f32_f32p(m, n, k, lhs, lhs_stride, rhs, dst, dst_stride, sizeof(float), clamp_min, clamp_max);
         } else if (fn_matmul_f16_f16p_f16p) {
             fn_matmul_f16_f16p_f16p(m, n, k, lhs, rhs, dst, dst_stride, sizeof(Float16), clamp_min, clamp_max);
         } else if (fn_matmul_f32_f32p_f32p) {
@@ -448,7 +445,7 @@ struct MatMulMethod {
                 m, n, k, reinterpret_cast<const uint16_t*>(lhs), rhs, reinterpret_cast<float*>(dst), dst_stride,
                 sizeof(float), clamp_min, clamp_max);
         } else if (fn_matmul_f16_bf16p_bf16p) {
-            fn_matmul_f16_bf16p_bf16p(m, n, k, lhs, rhs, dst, dst_stride, sizeof(__fp16), clamp_min, clamp_max);
+            fn_matmul_f16_bf16p_bf16p(m, n, k, lhs, rhs, dst, dst_stride, sizeof(uint16_t), clamp_min, clamp_max);
         } else {
             KAI_ERROR("Main kernel is not available!");
         }
